Updated plot_density.py
Changed logging to be debug statements, when -v is given it is in verbose mode and all debug statements will be active.

Added -i / --initial_position to include the initial position of the target atoms. color list is added to ensure the dots have the same color as the data of the same sim.

Updated plotting to have xlim from min(distance) - 1 > x > max(distance) + 1